+++ UCR/Views/Dialogs/StringDialogViewModel.cs
@@ -1,0 +1,9 @@
+﻿namespace HidWizards.UCR.Views.Dialogs
+{
+    public class StringDialogViewModel
+    {
+        public string Title { get; set; }
+        public string Hint { get; set; }
+        public string Value { get; set; }
+    }
+}